+++ tests/Propulsion.Kafka.Integration/ParallelIntegration.fs
@@ -9,7 +9,6 @@ open System
 open System.Collections.Concurrent
 open System.ComponentModel
 open System.Threading
-open System.Threading.Tasks
 open Xunit
 
 [<AutoOpen>]